refactor: use shared webpack post-create hook
